:recycle: :: infra계층에서 sendEmail 구현하기
