+++ goms-application/src/main/kotlin/com/goms/v2/domain/email/spi/EmailPort.kt
@@ -1,0 +1,7 @@
+package com.goms.v2.domain.email.spi
+
+interface EmailPort {
+
+    fun sendEmail(email: String, authKey: String)
+
+}
